refactor: move exception classees to exceptions package
Moved EmptyStringException, InvalidDateFormatException,
InvalidInputException, and RonaldoException from ui to exceptions
package. Also updated Storage.java and Ronaldo.java accordingly.
